+++ lib/flutter_animate_on_scroll.dart
@@ -5,4 +5,5 @@ export 'src/widgets/custom_animated_scroll.dart';
 export './src/widgets/fade/fade.dart';
 export './src/widgets/slide/slide.dart';
 export './src/widgets/rotate/rotate.dart';
-export './src/widgets/zoom/zoom_in.dart';
+export './src/widgets/zoom/zoom.dart';
+export './src/widgets/text/text.dart';