+++ lib/src/utils/utils.dart
@@ -5,10 +5,10 @@ final ScrollController _scrollController = ScrollController();
 
 extension BuildContextExtensions on BuildContext {
   /// get screen height using context directly
-  double get height => MediaQuery.of(this).size.height;
+  double get height => MediaQuery.sizeOf(this).height;
 
   /// get screen width using context directly
-  double get width => MediaQuery.of(this).size.width;
+  double get width => MediaQuery.sizeOf(this).width;
 
   /// get scrollcontroller using context directly
   ScrollController get scrollController => _scrollController;
@@ -26,16 +26,13 @@ extension NumDurationExtensions on num {
   Duration get seconds => Duration(microseconds: (this * 1000 * 1000).round());
 
   /// get minutes using num directly
-  Duration get minutes =>
-      Duration(microseconds: (this * 1000 * 1000 * 60).round());
+  Duration get minutes => Duration(microseconds: (this * 1000 * 1000 * 60).round());
 
   /// get hours using num directly
-  Duration get hours =>
-      Duration(microseconds: (this * 1000 * 1000 * 60 * 60).round());
+  Duration get hours => Duration(microseconds: (this * 1000 * 1000 * 60 * 60).round());
 
   /// get days using num directly
-  Duration get days =>
-      Duration(microseconds: (this * 1000 * 1000 * 60 * 60 * 24).round());
+  Duration get days => Duration(microseconds: (this * 1000 * 1000 * 60 * 60 * 24).round());
 }
 
 /// Shorthand to convert degrees to Radians. (multiply degrees with this value)
@@ -43,6 +40,3 @@ const double toRad = math.pi / 180.0;
 
 /// Shorthand to convert radians to Degrees. (multiply radians with this value)
 const double toDeg = 180.0 / math.pi;
-
-/// Percentage of visibility of widget
-const double visibility = 1.75;