+++ lib/src/widgets/rotate/rotate_in_down_right.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/scheduler.dart';
 
 import '../../utils/utils.dart';
@@ -24,6 +25,9 @@ class RotateInDownRight extends StatefulWidget {
   /// provide degree if need it, by default 45
   final double? degree;
 
+  /// provide repeated animation for widget, by default false
+  final bool? repeat;
+
   const RotateInDownRight({
     super.key,
     required this.child,
@@ -32,74 +36,105 @@ class RotateInDownRight extends StatefulWidget {
     this.duration,
     required this.globalKey,
     this.degree,
+    this.repeat = false,
   });
 
   @override
   State<RotateInDownRight> createState() => _RotateInDownRightState();
 }
 
-class _RotateInDownRightState extends State<RotateInDownRight>
-    with SingleTickerProviderStateMixin {
+class _RotateInDownRightState extends State<RotateInDownRight> with SingleTickerProviderStateMixin {
   late AnimationController _animationController;
   late Animation<double> _animation;
-  Offset _position = Offset.zero;
-  Size _size = const Size(0, 0);
-  bool _isAnimated = false;
+  final ValueNotifier<Offset> _position = ValueNotifier(Offset.zero);
+  final ValueNotifier<Size> _size = ValueNotifier(const Size(0, 0));
+  final ValueNotifier<bool> _isAnimated = ValueNotifier(false);
+  bool _isInView = false;
 
   @override
   void initState() {
+    super.initState();
+
     _animationController = AnimationController(
       vsync: this,
       duration: widget.duration ?? 300.ms,
     );
+
     _animation = Tween<double>(begin: 0.0, end: 1.0).animate(
       CurvedAnimation(
-          parent: _animationController,
-          curve: widget.curves ?? Curves.decelerate),
-    )..addListener(() {
-        setState(() {});
-      });
+        parent: _animationController,
+        curve: widget.curves ?? Curves.decelerate,
+      ),
+    );
+
     SchedulerBinding.instance.addPostFrameCallback((_) {
       if (widget.globalKey.currentContext == null) return;
-      RenderBox renderBox =
-          widget.globalKey.currentContext?.findRenderObject() as RenderBox;
+      RenderBox renderBox = widget.globalKey.currentContext!.findRenderObject() as RenderBox;
       Offset position = renderBox.localToGlobal(Offset.zero);
-      setState(() {
-        _position = position;
-        _size = renderBox.size;
-      });
-      if (context.height > _position.dy) {
-        _animate();
+      _position.value = position;
+      _size.value = renderBox.size;
+      if (context.height > _position.value.dy) {
+        _animate(isInView: true);
+        _isInView = true;
       }
     });
 
-    context.scrollController.addListener(_onScroll);
-    super.initState();
-  }
-
-  _animate() {
-    if (!mounted) return;
-    Future.delayed(widget.delay ?? 0.ms, () {
-      _animationController.forward(from: 0.0);
-      setState(() {
-        _isAnimated = true;
-      });
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      ScrollableState? scrollableState = Scrollable.of(context);
+      scrollableState.position.addListener(_onScroll);
     });
   }
 
-  _onScroll() {
-    if (_isAnimated) return;
-    if (context.scrollController.offset >
-        _position.dy - (_size.height * visibility) - (context.height / 2)) {
-      _animate();
+  void _animate({bool isScrollingUp = false, bool isInView = false}) {
+    if (!mounted) return;
+    Future.delayed(widget.delay ?? Duration.zero, () {
+      if (isScrollingUp && !isInView) {
+        _animationController.reverse(from: 1.0);
+      } else if (!isScrollingUp && isInView) {
+        _animationController.forward(from: 0.0);
+      }
+      _isAnimated.value = true;
+    });
+  }
+
+  void _onScroll() {
+    if (_isAnimated.value && widget.repeat == false) return;
+    ScrollableState? scrollableState = Scrollable.of(context);
+    final viewportDimension = scrollableState.position.viewportDimension;
+    final scrollPosition = scrollableState.position.pixels;
+    final widgetTop = _position.value.dy;
+    final widgetBottom = widgetTop + _size.value.height;
+
+    // check direction of scroll to animate
+    bool isScrollingDown = scrollableState.position.userScrollDirection == ScrollDirection.reverse;
+    bool isScrollingUp = scrollableState.position.userScrollDirection == ScrollDirection.forward;
+    // Check if the widget is within the viewport
+    bool isInView = scrollPosition < widgetBottom && (scrollPosition + viewportDimension) > widgetTop;
+
+    // Handle animation based on visibility and scroll direction
+    if (widget.repeat!) {
+      if (isInView && !_isInView && isScrollingDown) {
+        _animate(isScrollingUp: false, isInView: isInView);
+      } else if (isInView && !_isInView && isScrollingUp) {
+        _animate(isScrollingUp: true, isInView: isInView);
+      }
+    } else {
+      if (isInView && !_isInView) {
+        _animate(isInView: isInView);
+      }
     }
+
+    _isInView = isInView;
   }
 
   @override
   void dispose() {
+    ScrollableState? scrollableState = Scrollable.of(context);
+    scrollableState.position.removeListener(_onScroll);
     _animationController.dispose();
-    context.scrollController.removeListener(_onScroll);
-    context.scrollController.dispose();
+    _position.dispose();
+    _size.dispose();
+    _isAnimated.dispose();
     super.dispose();
   }