+++ lib/src/widgets/text/text.dart
@@ -1,0 +1,2 @@
+export 'text_typer.dart';
+export 'text_typer_wavy.dart';