+++ lib/src/widgets/text/text_typer.dart
@@ -1,19 +1,18 @@
 import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/scheduler.dart';
 
 import '../../utils/utils.dart';
 
-class FadeInDown extends StatefulWidget {
-  /// attach widget to animation child
-  final Widget child;
+class TextTyperAnimation extends StatefulWidget {
+  /// provide text to animate
+  final String text;
 
   /// provide delay duration if need it, by default zero
   final Duration? delay;
 
-  /// provide animation duration if need it, by default 300 milliseconds
+  /// provide animation duration if need it, by default 2 seconds
   final Duration? duration;
 
   /// provide animation curves if need it, by default [Curves.decelerate]
@@ -22,33 +21,50 @@ class FadeInDown extends StatefulWidget {
   /// require [GlobalKey] to get widget position and size
   final GlobalKey globalKey;
 
-  /// provide offset if need it, by default 50
-  final double? offset;
+  /// provide text style if need it, by default [Theme.of(context).textTheme.bodyMedium]
+  final TextStyle? textStyle;
+
+  /// provide text alignment if need it, by default [TextAlign.start]
+  final TextAlign? textAlign;
+
+  /// provide overflow if need it, by default [TextOverflow.clip]
+  final TextOverflow? overflow;
 
   /// provide repeated animation for widget, by default false
   final bool? repeat;
 
-  const FadeInDown({
+  /// provide fade option to fade the text that is not visible, by default false
+  final bool? fade;
+
+  /// provide opacity value to fade the text that is not visible, by default 0.5
+  final double? opacity;
+
+  const TextTyperAnimation({
     super.key,
-    required this.child,
+    required this.text,
     this.delay,
     this.curves,
     this.duration,
     required this.globalKey,
-    this.offset,
+    this.textStyle,
+    this.textAlign,
+    this.overflow,
     this.repeat = false,
+    this.fade = false,
+    this.opacity = 0.5,
   });
 
   @override
-  State<FadeInDown> createState() => _FadeInDownState();
+  State<TextTyperAnimation> createState() => _TextTyperAnimationState();
 }
 
-class _FadeInDownState extends State<FadeInDown> with SingleTickerProviderStateMixin {
+class _TextTyperAnimationState extends State<TextTyperAnimation> with SingleTickerProviderStateMixin {
   late AnimationController _animationController;
-  late Animation<double> _animation;
+  late Animation<int> _animation;
   final ValueNotifier<Offset> _position = ValueNotifier(Offset.zero);
   final ValueNotifier<Size> _size = ValueNotifier(const Size(0, 0));
   final ValueNotifier<bool> _isAnimated = ValueNotifier(false);
+  final ValueNotifier<int> _textLength = ValueNotifier(0);
   bool _isInView = false;
 
   @override
@@ -57,15 +73,13 @@ class _FadeInDownState extends State<FadeInDown> with SingleTickerProviderStateM
 
     _animationController = AnimationController(
       vsync: this,
-      duration: widget.duration ?? 300.ms,
+      duration: widget.duration ?? 2.seconds,
     );
 
-    _animation = Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(
-        parent: _animationController,
-        curve: widget.curves ?? Curves.decelerate,
-      ),
-    );
+    _animation = StepTween(begin: 0, end: widget.text.length).animate(_animationController)
+      ..addListener(() {
+        _textLength.value = _animation.value;
+      });
 
     SchedulerBinding.instance.addPostFrameCallback((_) {
       if (widget.globalKey.currentContext == null) return;
@@ -73,7 +87,7 @@ class _FadeInDownState extends State<FadeInDown> with SingleTickerProviderStateM
       Offset position = renderBox.localToGlobal(Offset.zero);
       _position.value = position;
       _size.value = renderBox.size;
-      if (context.height > _position.value.dy) {
+      if (MediaQuery.of(context).size.height > _position.value.dy) {
         _animate(isInView: true);
         _isInView = true;
       }
@@ -135,23 +149,42 @@ class _FadeInDownState extends State<FadeInDown> with SingleTickerProviderStateM
     _position.dispose();
     _size.dispose();
     _isAnimated.dispose();
+    _textLength.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
-      animation: _animation,
+      animation: _animationController,
       builder: (context, child) {
-        return Transform.translate(
-          offset: Offset(0, -(widget.offset ?? 50) * (1 - _animation.value)),
-          child: FadeTransition(
-            opacity: _animation,
-            child: Container(
-              key: widget.globalKey,
-              child: widget.child,
-            ),
-          ),
+        String visibleText = widget.text.substring(0, _textLength.value);
+        String remainingText = widget.text.substring(_textLength.value);
+        return Container(
+          key: widget.globalKey,
+          child: widget.fade == false
+              ? Text(
+                  visibleText,
+                  style: widget.textStyle ?? Theme.of(context).textTheme.bodyMedium,
+                  textAlign: widget.textAlign ?? TextAlign.start,
+                  overflow: widget.overflow ?? TextOverflow.clip,
+                )
+              : RichText(
+                  textAlign: widget.textAlign ?? TextAlign.start,
+                  overflow: widget.overflow ?? TextOverflow.clip,
+                  text: TextSpan(
+                    style: widget.textStyle ?? Theme.of(context).textTheme.bodyMedium,
+                    children: [
+                      TextSpan(text: visibleText), // Fully visible text
+                      TextSpan(
+                        text: remainingText,
+                        style: TextStyle(
+                          color: (widget.textStyle?.color ?? Colors.black).withOpacity(widget.opacity ?? 0.5),
+                        ),
+                      ), // Half transparent text
+                    ],
+                  ),
+                ),
         );
       },
     );