+++ CHANGELOG.md
@@ -10,3 +10,9 @@
 
 - Migrated from using `MediaQueryData.fromWindow` to `MediaQueryData.fromView` for improved accuracy in obtaining the viewport size.
 - Optimized the threshold calculation for triggering animations based on scroll position.
+
+## 0.1.0
+
+- Optimized scroll animation behaviour
+- Added ZoomOut animation
+- Added Text animation (typer effect with optional fade, wavy effect)